Created email package and moved email builder and dependencies into it
